+++ agent-inject/agent/container_sidecar_test.go
@@ -40,9 +40,11 @@ func TestContainerSidecarVolume(t *testing.T) {
 
 	pod := testPod(annotations)
 	var patches []*jsonpatch.JsonPatchOperation
-	agentConfig := AgentConfig{"foobar-image", "http://foobar:1234", DefaultVaultAuthType,
-		"test", "test", true, "1000", "100", DefaultAgentRunAsSameUser,
-		DefaultAgentSetSecurityContext, "", "map"}
+	agentConfig := AgentConfig{
+		"foobar-image", "http://foobar:1234", DefaultVaultAuthType, "test", "test", true, "1000", "100",
+		DefaultAgentRunAsSameUser, DefaultAgentSetSecurityContext, "", "map",
+		DefaultResourceRequestCPU, DefaultResourceRequestMem, DefaultResourceLimitCPU, DefaultResourceLimitMem,
+	}
 
 	err := Init(pod, agentConfig)
 	if err != nil {
@@ -105,11 +107,13 @@ func TestContainerSidecar(t *testing.T) {
 	pod := testPod(annotations)
 	var patches []*jsonpatch.JsonPatchOperation
 
-	agentConfig := AgentConfig{"foobar-image", "http://foobar:1234", DefaultVaultAuthType,
-		"test", "test", false, "1000", "100", DefaultAgentRunAsSameUser,
-		DefaultAgentSetSecurityContext, "http://proxy:3128", "map"}
+	agentConfig := AgentConfig{
+		"foobar-image", "http://foobar:1234", DefaultVaultAuthType, "test", "test", false, "1000", "100",
+		DefaultAgentRunAsSameUser, DefaultAgentSetSecurityContext, "https://proxy:3128", "map",
+		DefaultResourceRequestCPU, DefaultResourceRequestMem, DefaultResourceLimitCPU, DefaultResourceLimitMem,
+	}
 
-	err := Init(pod, agentConfig)
+  err := Init(pod, agentConfig)
 	if err != nil {
 		t.Errorf("got error, shouldn't have: %s", err)
 	}
@@ -222,11 +226,13 @@ func TestContainerSidecarRevokeHook(t *testing.T) {
 			pod := testPod(annotations)
 			var patches []*jsonpatch.JsonPatchOperation
 
-			agentConfig := AgentConfig{"foobar-image", "http://foobar:1234", DefaultVaultAuthType,
-				"test", "test", tt.revokeFlag, "1000", "100", DefaultAgentRunAsSameUser,
-				DefaultAgentSetSecurityContext, "", "map"}
+			agentConfig := AgentConfig{
+				"foobar-image", "http://foobar:1234", DefaultVaultAuthType, "test", "test", tt.revokeFlag, "1000", "100",
+				DefaultAgentRunAsSameUser, DefaultAgentSetSecurityContext, "", "map",
+				DefaultResourceRequestCPU, DefaultResourceRequestMem, DefaultResourceLimitCPU, DefaultResourceLimitMem,
+			}
 
-			err := Init(pod, agentConfig)
+      err := Init(pod, agentConfig)
 			if err != nil {
 				t.Errorf("got error, shouldn't have: %s", err)
 			}
@@ -275,11 +281,13 @@ func TestContainerSidecarConfigMap(t *testing.T) {
 	pod := testPod(annotations)
 	var patches []*jsonpatch.JsonPatchOperation
 
-	agentConfig := AgentConfig{"foobar-image", "http://foobar:1234", DefaultVaultAuthType,
-		"test", "test", true, "1000", "100", DefaultAgentRunAsSameUser,
-		DefaultAgentSetSecurityContext, "", "map"}
+	agentConfig := AgentConfig{
+		"foobar-image", "http://foobar:1234", DefaultVaultAuthType, "test", "test", true, "1000", "100",
+		DefaultAgentRunAsSameUser, DefaultAgentSetSecurityContext, "", "map",
+		DefaultResourceRequestCPU, DefaultResourceRequestMem, DefaultResourceLimitCPU, DefaultResourceLimitMem,
+	}
 
-	err := Init(pod, agentConfig)
+  err := Init(pod, agentConfig)
 	if err != nil {
 		t.Errorf("got error, shouldn't have: %s", err)
 	}
@@ -975,9 +983,11 @@ func TestContainerCache(t *testing.T) {
 			pod := testPod(tt.annotations)
 			var patches []*jsonpatch.JsonPatchOperation
 
-			agentConfig := AgentConfig{"foobar-image", "http://foobar:1234", DefaultVaultAuthType,
-				"test", "test", true, "1000", "100", DefaultAgentRunAsSameUser,
-				DefaultAgentSetSecurityContext, "", "map"}
+			agentConfig := AgentConfig{
+				"foobar-image", "http://foobar:1234", DefaultVaultAuthType, "test", "test", true, "1000", "100",
+				DefaultAgentRunAsSameUser, DefaultAgentSetSecurityContext, "", "map",
+				DefaultResourceRequestCPU, DefaultResourceRequestMem, DefaultResourceLimitCPU, DefaultResourceLimitMem,
+			}
 
 			err := Init(pod, agentConfig)
 			require.NoError(t, err)