+++ agent-inject/handler.go
@@ -49,6 +49,10 @@ type Handler struct {
 	SameID             bool
 	SetSecurityContext bool
 	DefaultTemplate    string
+	ResourceRequestCPU string
+	ResourceRequestMem string
+	ResourceLimitCPU   string
+	ResourceLimitMem   string
 }
 
 // Handle is the http.HandlerFunc implementation that actually handles the
@@ -156,6 +160,10 @@ func (h *Handler) Mutate(req *v1beta1.AdmissionRequest) *v1beta1.AdmissionRespon
 		SameID:             h.SameID,
 		SetSecurityContext: h.SetSecurityContext,
 		DefaultTemplate:    h.DefaultTemplate,
+		ResourceRequestCPU: h.ResourceRequestCPU,
+		ResourceRequestMem: h.ResourceRequestMem,
+		ResourceLimitCPU:   h.ResourceLimitCPU,
+		ResourceLimitMem:   h.ResourceLimitMem,
 	}
 	err = agent.Init(&pod, cfg)
 	if err != nil {